naoseimais

+++ action/setores.php
@@ -1,23 +1,25 @@
 <?php
-// include dos arquivos
-include_once   '../include/logado.php';
-include_once   '../include/conexao.php';
- 
-// captura a acao dos dados
-$acao = $_GET['acao'];
-$id = $_GET['id'];
- 
-// validacao
+
+include_once '../include/logado.php';
+include_once '../include/conexao.php';
+
+
+$acao = $_GET['acao'] ?? '';
+$id = $_GET['id'] ?? '';
+
+// executa a ação
 switch ($acao) {
     case 'excluir':
-        $sql = 'DELETE FROM setor WHERE SetorID = '.$id;
-        mysqli_query($conexao,$sql);
-        header('Location: ../lista-setores.php');
+        if (!empty($id)) {
+            $id = intval($id);
+            $sql = 'DELETE FROM setor WHERE SetorID = ' . $id;
+            mysqli_query($conexao, $sql);
+        }
+        header('Location: ../lista-setores.php'); 
         break;
-   
+
     default:
-        # code...
+        header('Location: ../lista-setores.php');
         break;
 }
 ?>
-
